+++ src/main/java/io/github/astrapi69/swing/tree/extension/DefaultMutableTreeNodeExtensions.java
@@ -22,9 +22,11 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing.tree.factory;
+package io.github.astrapi69.swing.tree.extension;
 
 import javax.swing.tree.DefaultMutableTreeNode;
+
+import io.github.astrapi69.swing.tree.factory.DefaultMutableTreeNodeFactory;
 
 /**
  * The class {@link DefaultMutableTreeNodeExtensions} provides methods for copy of