+++ src/main/java/io/github/astrapi69/swing/tree/factory/TreeNodeFactory.java
@@ -22,7 +22,7 @@
  * OF CONTRACT, TORT OR OTHERWISE, ARISING FROM, OUT OF OR IN CONNECTION
  * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
  */
-package io.github.astrapi69.swing.tree;
+package io.github.astrapi69.swing.tree.factory;
 
 import javax.swing.tree.DefaultMutableTreeNode;