+++ src/test/java/io/github/astrapi69/swing/tree/panel/DemoTreeNodeJTreeElementPanel.java
@@ -41,9 +41,9 @@ import io.github.astrapi69.swing.dialog.JOptionPaneExtensions;
 import io.github.astrapi69.swing.renderer.tree.GenericTreeElement;
 import io.github.astrapi69.swing.renderer.tree.JTreeElement;
 import io.github.astrapi69.swing.renderer.tree.renderer.JTreeNodeCellRenderer;
-import io.github.astrapi69.swing.tree.JTreeExtensions;
-import io.github.astrapi69.swing.tree.TreeNodeFactory;
+import io.github.astrapi69.swing.tree.extension.JTreeExtensions;
 import io.github.astrapi69.swing.tree.factory.DefaultMutableTreeNodeFactory;
+import io.github.astrapi69.swing.tree.factory.TreeNodeFactory;
 import io.github.astrapi69.swing.tree.panel.node.NodePanel;
 import io.github.astrapi69.test.object.Permission;