+++ src/test/java/io/github/astrapi69/swing/tree/panel/content/ApplicationPreferencesPanel.java
@@ -33,7 +33,7 @@ import javax.swing.tree.TreePath;
 
 import io.github.astrapi69.model.BaseModel;
 import io.github.astrapi69.model.api.IModel;
-import io.github.astrapi69.swing.tree.JTreeExtensions;
+import io.github.astrapi69.swing.tree.extension.JTreeExtensions;
 import io.github.astrapi69.swing.tree.model.TreeElement;
 import io.github.astrapi69.swing.tree.node.ComponentTreeNode;